`source': generally forbid during recursion..
Even though Heirloom mailx does allow this and seems to work fine
(differently to us) the general problem remains that the `source'd
file is placed onto the input stack and that won't be worked until
the compose mode has been left again and we return to the
mainloop.  E.g., it is completely broken.

Since we go this hard way we can also revert [6979c73] which was
added after an assertion i think, but of course it didn't address
the problem.  Who would spend a single thought that the original
implementor completely f....d the entire thing up _as such_?