Don't allow interruption during getaddrinfo(3) / gethostbyname(2)..
it seems the GNU/Linux GLibC getaddrinfo(3) resolver cannot deal
with interruption in that further calls won't work no more.
So make some love when the user tries nonetheless.

While here, do the same for gethostbyname(2) which should avoid
a possible leak, too.

P.S.: of course this is a no-no-no-go what we are doing in the
signal handler etc., but i want *this* functionality right now,
and it's up to the user if it is ok or not.  v15 will be correct,
too.