Add *mime-alternative-favour-rich*..
we won't be able to do selective MIME part printing until the v15
rewrite has taken place, but it is an annoyance that, e.g., many
newsletters produce a plain text part that is only a boiled down
reduced version of the HTML rich text one, so that you would have
to go to a web site in order to see what indeed is a vivid part
of the message, but hidden in some alternative part.

So what we can do the hacky and expensive way (an oh! and ach! on
c****y software design which leads to such hacks) is to implement
some lookahead if the user sets this new variable.
The only good thing on this variable is that we can keep it in
v15, though using some selective, explicit alternative sorting
(mutt(1) has something like that) will have to be provided, too.