FIX editline(3) compat (add n_ prefix to our TTY stuff)..
While testing after adding the new *history-** it turned out that
some editline(3) library ports are compiled so that the macro (?)
PROTECTED doesn't expand to static but to nothing, or whatever the
reason is, effectively resolving the tty_init() symbol from the
global namespace and thus our own NCL tty_init() symbol, causing
an infinite recursion and ultimately a program crash.  Sic.

So let's step at least this interface ahead to v15 and add a n_
prefix to all its functions, and, voilà, editline(3) works again.