Support `help' docstrings for `#', `-', plus..
While looking over the manual it became clear that the special
symbol commands were falsely sorted.  Thinking about that
i remembered that the sort order was because it was all temporary
and i wanted to mark out things which are no real commands from
the lex.c point of view (really?) and then i forgot about it.

So: introduce an array of these specials in order to be able to
print ?COMMAND documentation strings also for those.  Sort them
correctly in the manual.  Use NELEM() not a NULL name also for the
normal _cmd_tab[].