Be more friendly when composing mails (Martin Neitzel)..
by always printing To:, Cc:, Bcc: and Subject:, even if they are
still empty when the edits starts.
Realized after reading a mail of Martin on nail-devel@.

Of course this requires the usual hackery with prototype changes
in areas noone would think of, but that's just the way it is.
XXX The only real hack is that run_editor() was not changed but
XXX implicitly assumes compose mode when headers are given and
XXX an edit fd instead of a message structure.  This is the case
XXX and will remain for pre v15.
E.g., in v15 the user should be able to define and edit headers
freely whenever he desires.  Then all this has to change anyway.