nail.1: use display/show not type/print (Michael Convey)..
A thread on nail-devel@ sorted out that `type' replaced the
`print' command of research Unix mail' in 2BSD Mail, most likely
to avoid disambiguities in between teletype typing and printer
printing, keeping only a `p' alias.

For our manual let's now use display for header summaries and
message, well, display, and show for printing of general messages
etc., and drop all uses of printIng and typeIng unless it refers
to the actual commands.