Support :d specifier for at least `from' (Martin Neitzel)..
Martin Neitzel pointed out correctly on a cross-posting to
s-nail-users@ and nail-devel@ back on 2014-08-21 that the :d
modifier cannot be used to request a listing of deleted messages.
So hack in support for that, honoured by at least the `from'
command (in fact any using MSGLIST, but don't document it for that
terrible hack).