Add PS_SETFILE_OPENED (affects hooks), *bsdcompat*<>*emptystart*..
a thread on nail-devel@ made me realize that *bsdcompat* doesn't
affect messages regarding file open (*emptystart* is not BSD) and
when i looked into that disaster that we call setfile() it turned
out that a ternary return is not enough to differentiate in
between hard failure (-1), soft failure (1) and good (0) and have
mail (0) and empty mailbox (1).

Yes, we invoked hooks for other boxes than the one we tried to
open in the soft failure case, with *bsdcompat* / *bsdannounce* we
printed header summaries for the old box in that case, too,
possibly wiping away error messages from the screen.

Oh my god.
These words i first added in v14.6.
But: and they is still true!

Hack something in which avoids the complete no-goes (hook
invocation, announcements) by introducing the PS_SETFILE_OPENED
state bypass.

If we try to open a %USER box for a USER that doesn't exist, say
so.  And try to be a bit more *bsdcompat*ible by printing the
message that BSD users expect.  I hope i got the latter thing
right.

Finally, if we are about to open a nonexistent file but
*emptystart* set then don't let this cause hard failures which
would result in a program exit -- bend this and return soft
failure in this very case; use the BSD "no mail" message (see
[7abff87] (Fix "No mail for USER" message stuff for nonexistent
% box.., 2015-08-29)) if *bsdcompat* is set, print an error
message otherwise.

We should now handle *emptystart* / *bsdcompat* / empty file
/ nonexistent file correctly.
But the code is terrible.