Drop IS_TTY_SESSION(), pimp OPT_INTERACTIVE..
It doesn't make sense what we do.  Even though it may sometimes be
nice to be able to say

  $ s-nail | less

it is obvious that the above shouldn't count as an interactive
session, not only, but especially regarding our weak (sub)process
(group) and terminal handling.

Therefore simply require "STDIN=TTY + STDOUT=TTY == INTERACTIVE".

However, we should be much smarter regarding when we allow a PAGER
etc. to be used, which is supposed to be a possibly useful thing in

  $ s-nail -Scrt=0 >LOG 2>&1