(FIX) No $PAGER unless iactive or `[Mm]ore' / `[Pp]age', docu latter!..
Going through $PAGER can't be right except in non-interactive
mode, so ensure this.

While here i discovered that for this we have (one may still want
to do 'PAGER="cat -vet"' or the like) the `[Mm]ore' / `[Pp]age'
identicalos but which were either not documented correctly or not
at all!  Fix that!