cmd1.c: use sigman, needed for new colour..
the problem being that we want to do some cleanup and not simply
jump away on interrupt, which the code currently doesn't handle.
E.g., when interrupted within less(1) then less(1) ignores the
interrupt, but the code backing it (we, that is) does take care
for it.  This is completely brain damaged (like stated a thousand
times we should place kids in their own process group and isolate
them otherwise and when they take the terminal then it is THEIR
terminal and we shouldn't act upon just anything until they
return! DAMN!!) and now even results in crashes with the new
colour code, so ensure we do at least perform proper cleanup when
producing a brain fart.