FIX RFC2231 MIME param and do what comment says..
and perform a single n_iconv_str() on the completely joined parameter
string.  This should fix stateful encodings (which we do not truly
support, however).
Somehow it happened that the comment was correct but the iconv decoder
just run on a single decoded parameter continuation and not over the
final joined value.  I do not know wether some other MUAs only act like
that.