Move string dope from strings.c to the new memory.c..
Kurt Shoens introduced the "stringdope" in 2BSD (what i know
today).  I have extended it so that it plays much better with the
extended requirements that come with MIME etc., and am hacking for
years to get the string relaxation to all parts of this software,
to reduce memory pressure on the string dope.

But now that we drop the brilliant input stack idea in order to
get us going with recursion, the string dope will have to go, too.
As a first step, move it off strings.c where it stayed since at
least 2BSD, 1979-04 (36 years and seven months, at least) to the
new memory.c.