[privacy] Actually extend MIME parser for _SIGNED,_ENCRYPTED++..
Verify that _SIGNED/_ENCRYPTED have two bodies according to RFCs.
Extend send.c:_print_part_info() to include the corresponding
signature verification information.

And improve visual print/type experience by notifying that we've
decrypted something...

TODO   - A [Dd]ecryption operation should fail when decryption
TODO     fails.  We need atomic operations with rollback!

But we already want PGP support today, and we want a slightly
different message display, to be able to see the envelope that
carried the data for pk7m, for example, even if decryption is
possible that is.

P.S.:
=====

It doesn't work out like that, however, so reduce this one here to
only print something on PKCS7(M) decryption, which requires very
special and hacky things to be done (it is a wild jumping and
recursion that happens over here, just embed us where that needs
to be done).
Anything else has to wait until v15.0, just as written above.