mk-conf.sh: ignore further interrupts in traps..
when doing ^C^C or ditto otherwise during configuration we yet
would have left behind a possibly locked directory state ("$ make
distclean" being not available yet), especially since we did not
document nowhere what to do in this situation (as in, "$ rm
config.lst").

So ensure we do not honour interrupts when we are already
trap'ping, which should henceforth avoid such mess.