Drop WANT_EDITLINE support..
This topic branch will see the rewrite of the NCL to our new MLE
Mailx-Line-Editor, and that will be smaller and offer enough to
obsolete editline(3) support.

P.S.: in fact i think it is much smaller, more performant and
better, and i do like the tabulator expansion.  It needs some more
hands, like case-transformations etc., but that time will bring.