mk-conf.sh: add awareness of ENOSYS errors..
We yet used link_check(), but on Debian kFreeBSD we earn test failures.
I suspect that the GNU LibC monster provides ENOSYS stubs for whatever
function, i think it most likely is pipe2(2).

To circumvent that, switch some to run_check() and catch ENOSYS errors.
This will not help against booting a kernel that misses some functionality,
because at runtime we are still not prepared for that.
However, that is beyond of what most software does, and we are a primitive
MUA.  Still.  And always.  Nonetheless, for IPv6 support etc. runtime
should be the goal