Change make targets; add $MAILJOBS; don't track $DESTDIR (Gaetan Bisson)..
This is a follow-up to [f0bbd72] (makefile: place new
"_not_all__: build" first to avoid default "all".., 2015-12-08):

    i currently have a hard time looking at Debian build logs because
    the used sequence is "config [all] test install" and that results
    in three new evaluations of the configuration and rebuilds because
    that sequence should be "config build test doinstall" instead.
    This is hard violent pressure towards "./configure && make && make install"
    and i am all alone!

I have to admit that it was a fault right from the start, given
that we had "doinstall" and "packager-install" and such nonsense.
Thus and also finally honour a long-standing feature request of
Gaetan and don't track $DESTDIR in the configuration at all.
And change the targets so that the sequence now is

  $ make config build install
or
  $ make all install
or even
  $ make tangerine

but which also runs the test suite.
`install' (directly or indirectly) takes care for $DESTDIR.
While here allow parallelism for `build' and add $MAILJOBS which
may be set to, e.g., MAILJOBS='-j 2'.
Keep the `packager-install' target for backward compatibility.