Support *customhdr*, too (Sergey Matveev)..
Sergey correctly pointed out that `customhdr' is not covered by
copy-on-write environments, the successor to `localopts', and that
custom headers are a perfect subject to localization in accounts.

So add a *customhdr* variable that _can_ be placed under control
of `localopts', until the v15 framework hopefully allows
localization of things like alternates, aliases, shortcuts,
ghosts, and also custom headers.  Etc.

At that time it may become obsoleted again in favour of the
cleaner `customhdr', which will also save some code bytes.
Document as obsolete, therefore, and for now.