FIX MIME with *signature*s (Dirk-Wilhelm Peters)..
The content of *signature* files was yet simply embedded into the
final message, which Dirk-Wilhelm pointed out will potentially
break if the message will be send in a 7-bit character set.

In fact there was a long-standing in-code TODO entry (from
[6f633dc], 2013-02-09, thus three years less one week :) that
remarked that *signature* content should be embedded into the
message body, "thus" making it subject to the normal MIME
classification.  Just as is done for the "head" and "tail"
injection strings.