Fix SSL init order, change *ssl-rand-file* behaviour..
For one we have initialized the error strings after the SSL
library initialization.

Change behaviour of *ssl-rand-file*: if we cannot use this one,
use the SSL builtin path if, and only if, RAND_status() says that
the PRNG is not yet seeded.

Also, when loading RAND_load_file() is used and succeeds,
RAND_seed() in at least one chunk of homebrew "random" data before
saving the file.