send.c:sendpart(): "support" malformed messages..
It happens that mutt doesn't only support messages in base64
encoding that is false (silently ignoring the faulty data), but
also supports messages where the header and the body are not
separated with a(n) (completely) empty line, obviously silently
assuming that the body is what follows and doesn't specify a valid
header field.

That is sick, but simply not showing anything is also pretty
weird, as we did until now and another tested MUA does.
So do the same, but at least with error message, so that the user
knows (or can guess) we're guessing.

While here, join some useless locals into a single flag carrier.