Rewrite for a generic termcap(5) interface..
*term-ca-mode*, *line-editor-cursor-right*, this was all hacky
nonsense from the stone-age!

So do what the manual states since (nail.1: "Command line editor"
.. "On terminal control and line editor"..) and drop those,
replace them with *termcap* which allows the user to define
termcap(5) (later also: terminfo(5)) settings and
*termcap-disable* to disable interaction with the termcap library
even if that is available (turning *termcap* into sole provider).

The former NCL awaits the rewrite to be our new MLE, which then
works with the real termcap and also supports double-width
characters gracefully.  Until then (it will happen on [termcap.2],
too) simply remove the configurability of the former
*line-editor-cursor-right*.