main.c:_getopt(): slightly vaporise / tweak to allow -: option..
Our short option array doesn't start with colon : though comment
out all related code.

And then the loop was actually simple minded which caused the -:
not to be detected properly: we effectively skipped over any : in
the option array, likely thinking that it refers to a short option
before, meaning "takes an argument".
Instead check if we have an option match, if not check if the
current option takes an argument, and skip over two bytes in the
option array if this is the case.