n_str_dup(): don't assume NUL terminated source (Ralph Corderoy)..
Ralph reported a bound violation on the ML and pointed to
valgrind for detection purposes.  It turns out n_str_dup()
simply assumed a properly NUL terminated source string, but
that may not be true when we init struct str with a buffer,
instead the .l length field of the string is the thing to
look at, the terminator must be added in addition.