getapproval(): change prompt handling..
We have still leftovers from the Heirloom codebase which didn't
know about line editing and simply used "printf() + approval", but
that doesn't work if the user can ^A a.k.a. beginning-of-line,
since then the initial content will not be refreshed but be
subject of overwriting.  One more step is to unite the prefix into
the prompt.