Add n_msleep(), and use it..
We yet silently used the hammer sleep(3), which may even be
SIGALARM based, possibly, perhaps, or something like that.

Get rid of some TODOs and mk-conf.sh check for nanosleep(3), then
also check for sleep(3).  Provide a millisecond based function
which uses what we have found, and use that instead of sleep(3).

Note this won't work on old Solaris where nanosleep(2) requires
-lrt, and whereis we do link against that as necessary for MUA
itself we won't for the privilege-separated child.  This will be
changed in an upcoming commit in the topic/makefile-dev.2 branch.