Add n_strescsep() (allows backslash escaped separator)..
P.S.: during development of v14.9 it has become clear that we will
start being backward-incompatible, and support for sh(1)ell-style
argument quoting has been implemented on the [topic/wysh] branch.
There we've also added n_shell_sep(), which is the true successor
of strsep(), as it'll allow the user to specify _exactly_ what she
wants, and in a known and standardized syntax.  Of course much has
to change in the future to get there, but nonetheless: adding such
a fragile and non-standard (but for RFC 1524) syntax, like
n_strescsep(), is thus only a temporary solution.

Therefore usage of this function will be removed again as far as
possible during v14.9 development, and once only *customhdr*
remains (which is itself temporary), we'll make this function
local, to be removed in v15.