n_iconv_open(): use *charset-unknown-8bit* also for "binary"..
If we forcefully treat a MIME message part as text we
should really be prepared to convert a "binary" character
set to something useful, or at least not cause iconv(3)
error messages.

E.g., when sending a shell script or a roff file which embeds
control characters for field separation purposes, our MIME
classifier will switch to the "binary" charset unless it is
a well-known control character (like BEL), so this can indeed
happen easier than one might think.