makeprint(): filter some iswprint()==true, but which really aren't..
Ooooh, about the ISO C wide character interfaces, baby!  I waited
a bit with that, but today even lynx did something about it :)

Btw., with *headline-bidi*=3 the terminal i'm currently using
still looks better than mutt(1) with the newest ncurses when
displaying an arabic mail (in the summary), but of course it is
all a bad hack and the alignment isn't correct.
I hope my CText library, which will be used in v15, improves that
a bit.  And it would be so nice if we would _really_ support BIDI
as such, for our Arabic and Jewish cohabitants.