Introduce `wysh' command prefix to change argument parser..
Argument handling is a real mess.
E.g., we cannot have comments on a line except on a line by itself
because comments are not a part of the syntax but instead realized
as no-op commands.
Terribly inconsistent argument parsing extended over time with
things like laststring(), which rips off a mailbox/file name from
the end of the input string, thereafter the normal mechanism will
be used for the trimmed remains.

However: changing anything in here is backward incompatible.

So we cannot really do something about it until v15 crosses the
horizon and we can truly pulverize everybodies configuration!

Thus introduce a new "WhatYouSeeHell" command prefix that changes
parsing for the following command, to be somewhat sh(1)-style:

  ? mls @sdaoden\\.eu
  ? wysh mls '@sdaoden\.eu'
  ? wysh mls "@sdaoden\.eu"

To back this with code extend getrawlist() and optionally handle
a quite complete set of POSIX standardized sh(1)ell quoting rules
in there.
This isn't the final word, but it is a single step to let use be
used by users in a way that they know about from everydays shell
use.  But for the comment character.