FIX (fix?) `localopts'..
Now that we traverse up all the recursed macros and join all
to-be-unrolled variables in the outermost level (to avoid
redundantly setting or clearing them) one of my own macros didn't
work no more.

It turned out that i turn `localopts' on in a macro that performed
multiple account switches, which thus cleaned up settings which it
shouldn't.  Worse, it (thus) garbled settings of the `account'
cleanup list!

So terminate the list once we enter things like `account', even if
that means that the `localopts' of the outer macro isn't truly
honoured for anything that is done by the `account's.  This
behaviour, however, will be upward compatible since this is what
you would get from local environments, too.

P.S.: right now i don't dare to truly look wether the shipout
version can be fooled with that when one tries to do so.  We
actually did share the traverse lists always, but before had the
temporary_unroll hack, so i'm a little bit optimistic.  Also i am
**absolutely** sure i've tested the stuff heavily.  Well.

P.P.S.: i truly hope that [topic/shoens-2000], which effectively
will be rebased to *before* this changeset, will bring a bit of
healing for signals which happen to happen during
n_source_macro(), because otherwise the code as it is at the time
of this writing (from the [topic/okeys-attr.2] rewrite as such)
won't be safe regarding signals and i would have to reinstantiate
some kind of shitty temporary_unroll hack.  DAMN!  Sigh...