(BWDIC!) Redefine -u / $LOGNAME / $USER (Afan)..
Afan :=) (snailmail at ottenheimer dot com) unsuccessfully tried
to use -u in send mode as a replacement for setting $USER:

  I tested -u and got this error:

      The -u option cannot be used in send mode

And yes, we yet followed POSIX which only specifies -u as

  −u user  Read the system mailbox of the login name user. This
           shall only be successful if the invoking user has
           appropriate privileges to read the system mailbox of
           that user.

But the BSD Mail codebase always used this for more than that: it
assigns the given name to "myname" and uses that for quite a lot
of things, like MBOX From_ lines etc., i.e., it uses the "myname"
that normally is the user behind getuid(2)!
I remember i once sent a patch to NetBSD (and even OpenBSD?) to
straighten the handling to what S-nail changed to do, but i think
i didn't even get a response?  Our fork is the only one which does
so somewhat correctly, thus, and imho.  At least it verifies.

But thinking about it this is all wrong, and i think Afan
successfully pointed that out.

So redefine what we do:

  . "-u USER" is exactly the same as "-f %USER".

  . "myname" is no longer tainted by -u, and will be replaced with
    vlook(LOGNAME) on the long run.

  . upon startup we verify that $LOGNAME is set to the name of the
    active user, and correct as necessary, and complain, then.

  . shall $USER be set, we do the same.

  . $LOGNAME and $USER are now readonly.