Redefine *folder* / getfold->folder_query() / *HOME* (Justin Ellingwood)..
Accidentally i've (blush) searched the internet for "S-nail" and
found a tutorial written by Justin Ellingwood (JustinEllingwood at
gmail dot com) back in May (on Ubuntu/Postfix).  In there he writes

  You should get the following response:

    Output
    Can't canonicalize "/home/sammy/Maildir"

  This is normal and will only show during this first message.

Oh yes, unfortunately it is since we made *folder* a VIP and try
to resolve it (also via realpath(3)) whenever it is changed.  We
did so because of the problem of the BSD Mail / Heirloom mailx
codebase which cannot escape from a file/folder that is opened on
a relative path and that no longer resolved once we need to access
it again, resulting in a completely stuck codebase.  (The real
healing will be Void boxes again, but not before v15.)  Ah ja, and
then also because i don't want to see full path names in the
*prompt* and the plain `file'/`folder' when they reside under
*folder*: they should simply indicate that condition with the
leading "+" that is the syntax for this purpose.  These two
conditions wouldn't come together otherwise.

But now that we have our much improved variable flag handling we
can do a little bit better:

  - keep *folder* a VIP, but don't resolve it immediately once the
    variable is set, instead delay that until it is accessed the
    first time.

  - add a "private" (nodelete,readonly,undocumented) variable
    *-folder-resolved* that will keep track of that state.

  - since *folder* correlates to *HOME* handle that in the VIP
    handling.

  - drop the global "homedir" variable, go for vlook(HOME)
    instead (while here).

  - remove the global fio.c:var_folder_updated() function, merge
    that into the folder_query() that replaces the old getfold().

  - try to use realpath(, NULL) if possible: extend the mk-conf.sh
    test accordingly.

  - POSIX actually states that CWD is to be used when *folder* is
    not set or null.  Also do so if resolving failed.

This affects some callers: while here

  - drop some PATH_MAX stack buffers and alloca(3) uses.
    Go for our string dope instead.

  - in setfile() we yet may have already checked for wether
    a shortcut for "name" exists: if so, don't perform a second
    shortcut expansion in fexpand(), but call it with
    FEXP_NSHORTCUT and the expanded shortcut, instead.
    Also set FEXP_NSHELL always.

Plenty of room for improvements.
Error reporting.  Recursive directory creation.
And.And.And.