MLE: change HT expansion..
Adjust the MLE to look for and take care of PS_EXPAND_MULTIRESULT,
and dump that result array into a temporary file that can be
presented to the user via the $PAGER as necessary.

This is of course a hack, and it is intermediate: it would be nice
to have some numbers attached to each expansion and if the user
then inputs a number then that should be it.