Drop wordexp(3), instead unroll via fnmatch(3)..
Yes, some OSs don't have wordexp(3), or implement it with a sh(1)
subprocess, or even (Mac OS X Snow Leopard) perl(1) (there with
memory leaks, too).
We're better off like that.

While here add a PS_EXPAND_MULTIRESULT hack: it would be too
invasive for now to change fexpand() all the way to adjust the
return value for FEXP_MULTIOK, but now that HT expansion is
a regular feature we should improve it at least a bit so that we
don't inject potentially thousands of results onto the line.

So instead indicate through a global pstate hack that fexpand()
produced multiple results (only with FEXP_MULTIOK), and instead of
producing a single, ASCII SPACE separated string, generate an
array of NUL terminate strings that is terminated with NUL NUL,
then.