mk-conf.sh, mk-mk.in: diversify..
testing on old Solaris showed that we need -lrt for nanosleep(2)
there, and that is to say not only in the main MUA, but also in
the privilege-separated child: and that exceeds the capability of
our current makefile system!

I am too lazy to auto-generate the makefile via mk-conf.sh, and
implement program-specific rules as such, on the other hand the
current approach should not be used since libraries like OpenSSL,
Kerberos etc. should never be linked against the privilege-sepa-
rated child.

The quick-and-easy solution is to resort mk-conf.sh into "base"
and "extended functionality, and fork away macros like BASE_LIBS
etc. after the former have been worked, and to use those flags for
the sub-program(s) that we build instead of the fully fledged
stuff.  Since inference rules take no arguments, we need to
provide special make rules for the subprograms.

So later most functionality will be in a library, and then most
things have to change again anyway.  For now this is good enough.