Tweak VERBOSE handling (William Yodlowsky)..
Back in 2013 William asked for VERBOSE and pointed to a ML thread,
and wrote some statements.
It turns out i didn't realize what he was really saying, so throw
away all that if..echo(1)..fi statements that we use to realize
VERBOSEity, but instead use simple make(1) variable prefixes
instead.