FIX: no O_APPEND for .bz2/.xz/*file-hook-load*!  Use lseek(2)!!..
Do an lseek(2) on the file descriptor in order to truly adjust
the "file description" (standard wording.  No, it is not
a descriptor) offset when we will go over an external program to
load or save the mailbox.

So that change will eventually stop data corruption i have seen
with the hooks on OpenPGP encrypted files.
And: i wonder why i see that so late given... but, well.  DAMN!

Also, don't use ftruncate(2) but simply O_TRUNC.