Ensure reproducable environment..
Yet, if you configured with smake(1) but built with gmake(1),
then gmake(1) would be used for the job.  This is undesirable
in so far as it is not transparent.

Therefore generate a new file during configuration that holds
all the sh(1) and make(1) related settings, and source that
from within the top makefile, and use the read data already
when calling the (right) recursive make(1).